Moved Remote Image classes from test to separate file
