+++ EssentialFeed/EssentialFeedTests/Feed API/LoadImageCommentFromRemoteUseCaseTests.swift
@@ -10,74 +10,6 @@ import XCTest
 import Foundation
 import EssentialFeed
 
-class RemoteImageCommentLoader {
-	let url: URL
-	let httpClient: HTTPClient
-	
-	typealias Result = Swift.Result<[ImageComment], Error>
-	
-	enum Error: Swift.Error {
-		case connectivity
-		case invalidData
-	}
-	
-	init(url: URL, client: HTTPClient) {
-		self.url = url
-		self.httpClient = client
-	}
-	
-	func load(completion: @escaping (Result) -> Void) {
-		httpClient.get(from: url) { [weak self] (result) in
-			guard self != nil else { return }
-			
-			switch result {
-			case .failure:
-				completion(.failure(.connectivity))
-				
-			case let .success((data, response)):
-				completion(RemoteImageCommentLoader.map(data, response: response))
-			}
-		}
-	}
-	
-	static func map(_ data: Data, response: HTTPURLResponse) -> Result {
-		do {
-			let remoteComments = try ImageCommentMapper.map(data, from: response)
-			let comments = remoteComments.map { ImageComment(id: $0.id, message: $0.message, createdAt: $0.created_at, author: ImageCommentAuthor(username: $0.author.username))
-			}
-			
-			return .success(comments)
-		} catch {
-			return .failure(RemoteImageCommentLoader.Error.invalidData)
-		}
-	}
-}
-
-struct RemoteImageComment: Decodable {
-	let id: UUID
-	let message: String
-	let created_at: Date
-	let author: ImageCommentAuthor
-}
-
-class ImageCommentMapper {
-	private struct Root: Decodable {
-		let items: [RemoteImageComment]
-	}
-
-	static func map(_ data: Data, from response: HTTPURLResponse) throws -> [RemoteImageComment] {
-		let decoder = JSONDecoder()
-		decoder.dateDecodingStrategy = .iso8601
-		
-		guard response.statusCode == 200, let root = try? decoder.decode(Root.self, from: data) else {
-			throw RemoteImageCommentLoader.Error.invalidData
-		}
-		
-		return root.items
-	}
-}
-
-
 class LoadImageCommentFromRemoteUseCaseTests: XCTestCase {
 	func test_init_doesNotRequestDataOnInit() {
 		let (_, client) = makeSUT()
